+++ cookbooks/base/recipes/main.rb
@@ -22,10 +22,15 @@
 box = node[:box]
 
 # Language support
-package "language-pack-#{box['lang']}"
-package "language-pack-gnome-#{box['lang']}"
-package "aspell-#{box['lang']}"
-package "myspell-#{box['lang']}"
+package "language-pack-ca"
+package "language-pack-gnome-ca"
+package "aspell-ca"
+package "myspell-ca"
+
+package "language-pack-es"
+package "language-pack-gnome-es"
+package "aspell-es"
+package "myspell-es"
 
 
 # Gnome-Do: Quickly perform actions on your desktop
@@ -37,44 +42,5 @@ package "gnome-do-plugins"
 case platform
 when "mint"   then include_recipe "base::main-mint"
 when "ubuntu" then include_recipe "base::main-ubuntu"
-end
-
-
-# MS Office True Type Fonts
-fonts = data_bag_item('resources', 'fonts')
-msfonts = fonts['msttfonts']
-
-fonts_file   = msfonts['file']
-fonts_url    = msfonts['url']
-fonts_sha256 = msfonts['sha256']
-
-cache_path = Chef::Config[:file_cache_path]
-fonts_path = "/usr/share/fonts/truetype/msttfonts"
-
-remote_file "#{cache_path}/#{fonts_file}" do
-  source fonts_url
-  checksum fonts_sha256
-  notifies :run, "bash[install_fonts]", :immediately
-end
-
-bash "install_fonts" do
-  cwd cache_path
-  code <<-EOH
-    [[ ! -d "#{fonts_path}" ]] && mkdir -m 755 -p #{fonts_path}
-    tar -C #{fonts_path} -xzf #{fonts_file} \
-    && chown root.root #{fonts_path}/* \
-    && chmod 644 #{fonts_path}/*
-    EOH
-  action :nothing
-end
-
-box['users'].each do |username, usr|
-  execute "#{username}-load_fonts" do
-    command "fc-cache -fv"
-    user username
-    group usr['group']
-    action :nothing
-    subscribes :run, resources("bash[install_fonts]"), :immediately
-  end
 end